DMC changes

+++ lib/playlist/playlist_m3u.c
@@ -19,7 +19,7 @@ int ret;
     line.len -= 1;
     if(!str_diffn(line.s, "#EXTM3U", 7)) {
     } else if(!str_diffn(line.s, "#EXTINF", 7)) {
-      size_t len;
+      unsigned long len;
       size_t index = 8;
       index += scan_ulong(&line.s[index], &len);
       entry.length = len;
